Refactor view controllers
Make a IncomeCalculator class to have all model logics
Factor out redundant codes in SalaryViewController and HourlyWageViewController into a single parent view controller class
